# core/vm: Move interprepreter intrerruption check to jump instructions