Fixes https://github.com/ethereum/go-ethereum/issues/23969