@gumb0 2021-12-01
On my machine all 3 variants (no check, check every jump, check every 64th jump) look to have the same performance.

No check vs check every jump:
```
name                                  old time/op  new time/op  delta
main/blake2b_shifts/2805nulls         13.2ms ± 2%  13.1ms ± 2%  -0.88%  (p=0.004 n=16+17)
main/blake2b_shifts/5610nulls         29.9ms ± 1%  29.7ms ± 1%  -0.51%  (p=0.003 n=19+20)
main/blake2b_shifts/8415nulls         44.0ms ± 0%  43.8ms ± 1%  -0.59%  (p=0.000 n=18+18)
main/weierstrudel/0                    470µs ± 1%   469µs ± 1%    ~     (p=0.513 n=20+19)
main/weierstrudel/1                    978µs ± 0%   984µs ± 1%  +0.68%  (p=0.000 n=20+18)
main/weierstrudel/3                   1.52ms ± 0%  1.54ms ± 0%  +1.27%  (p=0.000 n=18+20)
main/weierstrudel/9                   3.12ms ± 1%  3.16ms ± 1%  +1.58%  (p=0.000 n=20+20)
main/weierstrudel/14                  4.45ms ± 1%  4.51ms ± 0%  +1.38%  (p=0.000 n=18+17)
main/sha1_divs/empty                   194µs ± 1%   194µs ± 0%    ~     (p=0.723 n=20+19)
main/sha1_divs/1351                   3.79ms ± 0%  3.79ms ± 0%    ~     (p=0.773 n=19+19)
main/sha1_divs/2737                   7.38ms ± 1%  7.38ms ± 0%    ~     (p=0.987 n=16+19)
main/sha1_divs/5311                   14.6ms ± 4%  14.5ms ± 1%    ~     (p=0.515 n=18+20)
main/sha1_shifts/empty                 158µs ± 1%   152µs ± 1%  -3.27%  (p=0.000 n=18+19)
main/sha1_shifts/1351                 3.13ms ± 1%  3.03ms ± 1%  -3.24%  (p=0.000 n=20+19)
main/sha1_shifts/2737                 6.06ms ± 0%  5.87ms ± 1%  -3.11%  (p=0.000 n=18+18)
main/sha1_shifts/5311                 11.9ms ± 0%  11.5ms ± 1%  -3.21%  (p=0.000 n=16+18)
main/blake2b_huff/empty                102µs ± 0%   103µs ± 1%  +0.69%  (p=0.000 n=18+20)
main/blake2b_huff/2805nulls           1.94ms ± 2%  1.94ms ± 1%    ~     (p=0.457 n=20+20)
main/blake2b_huff/5610nulls           3.73ms ± 2%  3.73ms ± 1%    ~     (p=0.443 n=18+19)
main/blake2b_huff/8415nulls           5.48ms ± 1%  5.46ms ± 1%    ~     (p=0.331 n=20+18)
micro/JUMPDEST_n0/empty               2.49ms ± 1%  2.49ms ± 1%  +0.26%  (p=0.004 n=20+20)
micro/jump_around/empty                144µs ± 0%   144µs ± 1%    ~     (p=0.330 n=18+16)
micro/memory_grow_mstore/nogrow        440µs ± 1%   440µs ± 0%    ~     (p=0.781 n=19+16)
micro/memory_grow_mstore/by1           467µs ± 1%   466µs ± 1%    ~     (p=0.101 n=19+20)
micro/memory_grow_mstore/by16          576µs ± 1%   576µs ± 1%    ~     (p=0.627 n=16+18)
micro/memory_grow_mstore/by32          655µs ± 0%   656µs ± 1%    ~     (p=0.258 n=20+19)
micro/signextend/zero                  318µs ± 1%   322µs ± 2%  +1.50%  (p=0.000 n=19+20)
micro/signextend/one                   328µs ± 1%   331µs ± 1%  +0.95%  (p=0.000 n=20+20)
micro/loop_with_many_jumpdests/empty  1.84µs ± 3%  1.85µs ± 4%    ~     (p=0.583 n=20+19)
micro/memory_grow_mload/nogrow         214µs ± 1%   215µs ± 1%  +0.36%  (p=0.010 n=20+18)
micro/memory_grow_mload/by1            230µs ± 1%   231µs ± 1%  +0.56%  (p=0.001 n=17+19)
micro/memory_grow_mload/by16           309µs ± 5%   317µs ± 4%  +2.62%  (p=0.005 n=19+18)
micro/memory_grow_mload/by32           412µs ± 3%   410µs ± 2%    ~     (p=0.056 n=17+20)
[Geo mean]                            1.13ms       1.13ms       -0.14%
```

Check every jump vs every 64th jump:
```
name                                  old time/op  new time/op  delta
main/blake2b_shifts/2805nulls         13.2ms ± 2%  13.2ms ± 4%    ~     (p=0.765 n=16+20)
main/blake2b_shifts/5610nulls         29.9ms ± 1%  29.1ms ± 2%  -2.40%  (p=0.000 n=19+15)
main/blake2b_shifts/8415nulls         44.0ms ± 0%  43.0ms ± 1%  -2.29%  (p=0.000 n=18+19)
main/weierstrudel/0                    470µs ± 1%   460µs ± 1%  -2.12%  (p=0.000 n=20+19)
main/weierstrudel/1                    978µs ± 0%   947µs ± 0%  -3.16%  (p=0.000 n=20+18)
main/weierstrudel/3                   1.52ms ± 0%  1.47ms ± 0%  -3.21%  (p=0.000 n=18+19)
main/weierstrudel/9                   3.12ms ± 1%  3.03ms ± 0%  -2.83%  (p=0.000 n=20+17)
main/weierstrudel/14                  4.45ms ± 1%  4.32ms ± 1%  -2.95%  (p=0.000 n=18+20)
main/sha1_divs/empty                   194µs ± 1%   194µs ± 1%    ~     (p=0.119 n=20+18)
main/sha1_divs/1351                   3.79ms ± 0%  3.83ms ± 1%  +1.11%  (p=0.000 n=19+20)
main/sha1_divs/2737                   7.38ms ± 1%  7.46ms ± 1%  +1.14%  (p=0.000 n=16+18)
main/sha1_divs/5311                   14.6ms ± 4%  14.6ms ± 1%    ~     (p=0.361 n=18+20)
main/sha1_shifts/empty                 158µs ± 1%   155µs ± 1%  -1.87%  (p=0.000 n=18+20)
main/sha1_shifts/1351                 3.13ms ± 1%  3.07ms ± 1%  -1.78%  (p=0.000 n=20+20)
main/sha1_shifts/2737                 6.06ms ± 0%  5.95ms ± 1%  -1.76%  (p=0.000 n=18+17)
main/sha1_shifts/5311                 11.9ms ± 0%  11.7ms ± 1%  -2.08%  (p=0.000 n=16+20)
main/blake2b_huff/empty                102µs ± 0%   101µs ± 0%  -1.26%  (p=0.000 n=18+18)
main/blake2b_huff/2805nulls           1.94ms ± 2%  1.91ms ± 1%  -1.25%  (p=0.000 n=20+19)
main/blake2b_huff/5610nulls           3.73ms ± 2%  3.68ms ± 1%  -1.28%  (p=0.000 n=18+19)
main/blake2b_huff/8415nulls           5.48ms ± 1%  5.40ms ± 1%  -1.42%  (p=0.000 n=20+18)
micro/JUMPDEST_n0/empty               2.49ms ± 1%  2.45ms ± 1%  -1.48%  (p=0.000 n=20+20)
micro/jump_around/empty                144µs ± 0%   142µs ± 1%  -1.32%  (p=0.000 n=18+20)
micro/memory_grow_mstore/nogrow        440µs ± 1%   433µs ± 1%  -1.78%  (p=0.000 n=19+20)
micro/memory_grow_mstore/by1           467µs ± 1%   458µs ± 1%  -1.94%  (p=0.000 n=19+20)
micro/memory_grow_mstore/by16          576µs ± 1%   569µs ± 2%  -1.22%  (p=0.000 n=16+19)
micro/memory_grow_mstore/by32          655µs ± 0%   660µs ± 0%  +0.65%  (p=0.000 n=20+16)
micro/signextend/zero                  318µs ± 1%   344µs ± 1%  +8.20%  (p=0.000 n=19+18)
micro/signextend/one                   328µs ± 1%   355µs ± 1%  +8.04%  (p=0.000 n=20+19)
micro/loop_with_many_jumpdests/empty  1.84µs ± 3%  1.83µs ± 2%    ~     (p=0.129 n=20+20)
micro/memory_grow_mload/nogrow         214µs ± 1%   211µs ± 1%  -1.39%  (p=0.000 n=20+19)
micro/memory_grow_mload/by1            230µs ± 1%   226µs ± 1%  -1.43%  (p=0.000 n=17+19)
micro/memory_grow_mload/by16           309µs ± 5%   312µs ± 3%    ~     (p=0.066 n=19+17)
micro/memory_grow_mload/by32           412µs ± 3%   411µs ± 4%    ~     (p=0.832 n=17+18)
[Geo mean]                            1.13ms       1.13ms       -0.71%
```