@chfast 2021-12-01
I think we need some unit tests for this use case first: execute long running loop in one go-routine and abort it in second go-routine.