@holiman 2021-12-01
If it's indeed 'natively' atomic on `amd64`, IMO we could just check every jump, and not bother with the counter