@chfast 2021-12-01
> If it's indeed 'natively' atomic on `amd64`, IMO we could just check every jump, and not bother with the counter

It is just `MOV` instruction because on `amd64` every memory load is atomic by design. But this is not the case on `arm64` so someone with raspberry pi or Apple M1 may want to benchmark this.
So with the counter we have 1 load + 1 store to just handle the counter. Without counter we just have 1 load.

However, the version with counter looks faster on average, but not in a benchmark purely focusing on jumps like "jump_around".

Interestingly, in the `disasm` view in the version without counter the `errStopToken` branch is considered unlikely and move to the end of the function (good decision). This does not happen in the version with the counter. I was not able to reproduce this compiler decision in isolation.

```
name                                  old time/op  new time/op  delta                                                                                                                                                             
micro/memory_grow_mload/nogrow         221µs ± 1%   220µs ± 0%    ~     (p=0.089 n=20+17)                                                                                                                                         
micro/memory_grow_mload/by1            230µs ± 1%   230µs ± 1%    ~     (p=0.624 n=19+19)                                                                                                                                         
micro/memory_grow_mload/by16           291µs ± 1%   292µs ± 1%    ~     (p=0.534 n=20+20)                                                                                                                                         
micro/memory_grow_mload/by32           366µs ± 1%   367µs ± 1%  +0.25%  (p=0.041 n=19+20)                                                                                                                                         
micro/signextend/zero                  407µs ± 0%   399µs ± 0%  -2.17%  (p=0.000 n=20+20)                                                                                                                                         
micro/signextend/one                   422µs ± 1%   409µs ± 1%  -3.11%  (p=0.000 n=20+20)                                                                                                                                         
micro/loop_with_many_jumpdests/empty  1.70µs ± 2%  1.69µs ± 2%    ~     (p=0.441 n=20+20)                                                                                                                                         
micro/jump_around/empty                146µs ± 1%   146µs ± 1%    ~     (p=0.141 n=20+19)                                                                                                                                         
micro/memory_grow_mstore/nogrow        454µs ± 1%   453µs ± 1%  -0.23%  (p=0.001 n=18+20)                                                                                                                                         
micro/memory_grow_mstore/by1           468µs ± 0%   468µs ± 1%    ~     (p=0.698 n=20+20)                                                                                                                                         
micro/memory_grow_mstore/by16          548µs ± 0%   549µs ± 1%  +0.30%  (p=0.000 n=18+20)                                                                                                                                         
micro/memory_grow_mstore/by32          612µs ± 0%   612µs ± 1%    ~     (p=0.620 n=20+20)                                                                                                                                         
micro/JUMPDEST_n0/empty               2.84ms ± 0%  2.85ms ± 0%  +0.07%  (p=0.028 n=16+17)                                                                                                                                         
main/sha1_divs/empty                   204µs ± 0%   197µs ± 0%  -3.60%  (p=0.000 n=17+20)
main/sha1_divs/1351                   4.14ms ± 0%  3.98ms ± 0%  -3.85%  (p=0.000 n=17+19)
main/sha1_divs/2737                   8.08ms ± 0%  7.76ms ± 0%  -3.97%  (p=0.000 n=18+19)
main/sha1_divs/5311                   15.8ms ± 0%  15.2ms ± 0%  -3.81%  (p=0.000 n=20+18)
main/weierstrudel/0                    468µs ± 0%   467µs ± 1%  -0.19%  (p=0.002 n=19+19)
main/weierstrudel/1                    978µs ± 1%   976µs ± 0%  -0.26%  (p=0.004 n=19+19)
main/weierstrudel/3                   1.54ms ± 1%  1.54ms ± 1%    ~     (p=0.184 n=19+20)
main/weierstrudel/9                   3.21ms ± 1%  3.19ms ± 1%  -0.67%  (p=0.000 n=20+19)
main/weierstrudel/14                  4.61ms ± 1%  4.56ms ± 1%  -0.96%  (p=0.000 n=20+19)
main/sha1_shifts/empty                 151µs ± 1%   148µs ± 1%  -1.55%  (p=0.000 n=20+19)
main/sha1_shifts/1351                 3.09ms ± 1%  3.02ms ± 1%  -2.41%  (p=0.000 n=19+19)
main/sha1_shifts/2737                 6.03ms ± 0%  5.88ms ± 0%  -2.37%  (p=0.000 n=17+18)
main/sha1_shifts/5311                 11.8ms ± 0%  11.5ms ± 0%  -2.26%  (p=0.000 n=20+16)
main/blake2b_huff/empty                102µs ± 1%   100µs ± 0%  -2.18%  (p=0.000 n=19+18)
main/blake2b_huff/2805nulls           1.96ms ± 0%  1.90ms ± 1%  -2.88%  (p=0.000 n=19+20)
main/blake2b_huff/5610nulls           3.80ms ± 1%  3.71ms ± 1%  -2.50%  (p=0.000 n=20+20)
main/blake2b_huff/8415nulls           5.57ms ± 1%  5.42ms ± 0%  -2.65%  (p=0.000 n=16+18)
main/blake2b_shifts/2805nulls         14.4ms ± 1%  14.3ms ± 1%  -0.60%  (p=0.000 n=20+18)
main/blake2b_shifts/5610nulls         28.7ms ± 1%  28.6ms ± 1%  -0.40%  (p=0.012 n=20+20)
main/blake2b_shifts/8415nulls         42.7ms ± 0%  42.7ms ± 2%    ~     (p=0.626 n=16+20)
[Geo mean]                            1.16ms       1.14ms       -1.29%
```